requirements.txt
fastapi
uvicorn
python-dotenv
openai-agents
pydantic
tensorflow
requests